# src/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "2팀-How Do I Look",
    "description": "이 프로젝트는 코드잇 Node.js 백엔드 스프린트의 초급 팀 프로젝트입니다.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/styles/ranking": {
      "get": {
        "tags": ["스타일"],
        "summary": "스타일 랭킹 조회",
        "description": "- 랭킹을 기준으로 스타일을 조회하는 API입니다.\n- `sort`, `page`, `pageSize` 파라미터를 통해 랭킹 정렬 방식, 페이지 번호, 페이지당 항목 수를 조정할 수 있습니다.",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "스타일 랭킹을 정렬하는 기준. 예: 'viewCount', 'curationCount' 등"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "조회할 페이지 번호. 기본값은 1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "페이지당 스타일 항목 수. 기본값은 10"
          }
        ],
        "responses": {
          "200": {
            "description": "OK (성공)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "currentPage": {
                      "type": "integer",
                      "description": "현재 페이지 번호"
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "전체 페이지 수"
                    },
                    "totalItems": {
                      "type": "integer",
                      "description": "전체 스타일 항목 수"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "스타일의 고유 ID"
                          },
                          "nickname": {
                            "type": "string",
                            "description": "스타일을 등록한 사용자의 닉네임"
                          },
                          "title": {
                            "type": "string",
                            "description": "스타일의 제목"
                          },
                          "viewCount": {
                            "type": "integer",
                            "description": "스타일의 조회수"
                          },
                          "curationCount": {
                            "type": "integer",
                            "description": "스타일에 대한 큐레이팅 수"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "스타일의 생성 일시"
                          }
                        }
                      },
                      "description": "스타일 목록"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/styles/{styleId}/curations": {
      "get": {
        "tags": ["큐레이션"],
        "summary": "큐레이션 목록 조회",
        "description": "- 스타일을 조회할 경우 그 스타일에 해당되는 큐레이션 목록이 같이 조회됩니다.\n- 각 큐레이션의 트렌디, 개성, 실용성, 가성비 점수와 한줄 큐레이팅, 닉네임이 표시됩니다.\n- 닉네임, 내용으로 검색이 가능합니다.\n- 큐레이팅에 남겨진 답글도 같이 조회됩니다.",
        "parameters": [
          {
            "name": "styleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "스타일 ID (0 이상의 정수)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "현재 페이지 번호 (1 이상의 정수)"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "페이지당 아이템 수 (1 이상의 정수)"
          },
          {
            "name": "searchBy",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "검색 기준 (nickname | content)"
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "검색어"
          }
        ],
        "responses": {
          "200": {
            "description": "OK (성공)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "currentPage": {
                      "type": "integer",
                      "description": "현재 페이지 번호",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "전체 페이지 수",
                      "example": 10
                    },
                    "totalItemCount": {
                      "type": "integer",
                      "description": "전체 아이템 수",
                      "example": 100
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "큐레이션 ID",
                            "example": 1
                          },
                          "nickname": {
                            "type": "string",
                            "description": "닉네임",
                            "example": "큐레이션 작성자"
                          },
                          "content": {
                            "type": "string",
                            "description": "내용",
                            "example": "큐레이션 작성자"
                          },
                          "trendy": {
                            "type": "integer",
                            "description": "트렌디 점수",
                            "example": 10
                          },
                          "personality": {
                            "type": "integer",
                            "description": "개성 점수",
                            "example": 7
                          },
                          "practicality": {
                            "type": "integer",
                            "description": "실용성 점수",
                            "example": 9
                          },
                          "costEffectiveness": {
                            "type": "integer",
                            "description": "비용 효율성 점수",
                            "example": 8
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "생성 일시",
                            "example": "2025-03-10T09:30:00Z"
                          },
                          "comment": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "description": "댓글 ID",
                                "example": 2
                              },
                              "nickname": {
                                "type": "string",
                                "description": "스타일 및 댓글 작성자의 닉네임",
                                "example": "원작성자"
                              },
                              "content": {
                                "type": "string",
                                "description": "댓글 내용",
                                "example": "평가 고맙습니다"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time",
                                "description": "댓글 생성 일시",
                                "example": "2025-03-10T09:30:00Z"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "잘못된 요청입니다"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["큐레이션"],
        "summary": "큐레이션 등록",
        "description": "- 새로운 큐레이션을 등록하는 API입니다.\n- `styleId`는 필수로 입력되어야 합니다.\n- 요청 본문에 `nickname`, `content`, `password`, 점수 관련 항목을 포함해야 합니다.",
        "parameters": [
          {
            "name": "styleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "스타일 ID (0 이상의 정수)"
          }
        ],
        "requestBody": {
          "description": "비밀번호는 영문, 숫자 조합 8~16자",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "description": "큐레이션을 등록하는 사용자의 닉네임"
                  },
                  "content": {
                    "type": "string",
                    "description": "큐레이션의 본문 내용"
                  },
                  "password": {
                    "type": "string",
                    "description": "큐레이션 작성 시의 비밀번호",
                    "example": "password911"
                  },
                  "trendy": {
                    "type": "integer",
                    "description": "큐레이션의 트렌디 점수 (0~10 사이의 정수)"
                  },
                  "personality": {
                    "type": "integer",
                    "description": "큐레이션의 개성 점수 (0~10 사이의 정수)"
                  },
                  "practicality": {
                    "type": "integer",
                    "description": "큐레이션의 실용성 점수 (0~10 사이의 정수)"
                  },
                  "costEffectiveness": {
                    "type": "integer",
                    "description": "큐레이션의 비용 효율성 점수 ((0~10 사이의 정수)"
                  }
                },
                "required": [
                  "nickname",
                  "content",
                  "password",
                  "trendy",
                  "personality",
                  "practicality",
                  "costEffectiveness"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK (성공)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "description": "큐레이션의 고유 ID",
                      "example": 123
                    },
                    "nickname": {
                      "type": "string",
                      "description": "큐레이션을 등록한 사용자의 닉네임"
                    },
                    "content": {
                      "type": "string",
                      "description": "큐레이션의 본문 내용"
                    },
                    "trendy": {
                      "type": "integer",
                      "description": "큐레이션의 트렌디 점수"
                    },
                    "personality": {
                      "type": "integer",
                      "description": "큐레이션의 개성 점수"
                    },
                    "practicality": {
                      "type": "integer",
                      "description": "큐레이션의 실용성 점수"
                    },
                    "costEffectiveness": {
                      "type": "integer",
                      "description": "큐레이션의 비용 효율성 점수"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "큐레이션의 생성 일시"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "잘못된 요청입니다",
                      "example": "잘못된 요청입니다"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}